ui: add utilities to tint platform Switch
Shares most code with SwitchCompat but is not interface-compatible
